pagina de novos escritores 100%

+++ src/socket/deleteTopic.js
@@ -1,26 +1,16 @@
-module.exports = {
-    async index(server) {
-        const { Server, Socket } = require('socket.io');
+function deleteTopic(socket) {
+    socket.on('deleteTopic', async data => {
+        const SearchAdmin = require('../models/SearchAdmin');
+        const res = await SearchAdmin('user', data.user);
 
-        let options = {
-            pingTimeout: 30000,
+        if (res !== true && res[0].admin === 'true' && res[0].admin_level >= 2) {
+            const deleteTopic = require('../models/deleteTopic');
+            await deleteTopic(data.lang);
+            socket.emit('resTopic', { res: true });
+        } else {
+            socket.emit('resTopic', { res: false });
         }
-        const io = new Server(server, options)
-
-        io.on("connection", socket => {
-            console.log('new conn: ', socket.id);
-            socket.on('deleteTopic', async data => {
-                const SearchUser = require('../models/SearchUser');
-                const res = await SearchUser('username', data.user);
-
-                if (res[0].admin === 'true' && res[0].admin_level >= 2) {
-                    const deleteTopic = require('../models/deleteTopic');
-                    await deleteTopic(data.lang);
-                    socket.emit('resTopic', {res: true});
-                } else {
-                    socket.emit('resTopic', {res: false});
-                }
-            })
-        });
-    }
+    })
 }
+
+module.exports = deleteTopic
